leafdetect-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pillow==11.3.0
sqlalchemy==2.0.43
